refactor: separate component for new form
